refactor: extract user permission management to standalone component
- Create UserPermissionManagement component in /components
- Extract 415 lines from admin-management-interface.tsx to separate component
- Component features:
  * Fetch users with include_permissions=true
  * Inline college_code editing (super_admin only)
  * Bulk scholarship assignment modal
  * Uses new bulk API endpoints
- Replace users tab content with <UserPermissionManagement />
- Delete redundant /app/admin/user-management/page.tsx
- Improves maintainability and code organization
