+++ frontend/components/admin-management-interface.tsx
@@ -37,6 +37,7 @@ import {
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Textarea } from "@/components/ui/textarea";
 import { UserEditModal } from "@/components/user-edit-modal";
+import { UserPermissionManagement } from "@/components/user-permission-management";
 import apiClient, {
   AnnouncementCreate,
   AnnouncementUpdate,
@@ -2159,419 +2160,7 @@ export function AdminManagementInterface({
         </TabsContent>
 
         <TabsContent value="users" className="space-y-4">
-          <div className="flex items-center justify-between">
-            <h3 className="text-lg font-semibold">使用者權限管理</h3>
-            <div className="flex gap-2">
-              <Button
-                onClick={() => setShowUserForm(true)}
-                className="nycu-gradient text-white"
-              >
-                <Plus className="h-4 w-4 mr-1" />
-                新增使用者權限
-              </Button>
-              <Button variant="outline">
-                <Upload className="h-4 w-4 mr-1" />
-                批次匯入
-              </Button>
-            </div>
-          </div>
-
-          {/* 使用者統計卡片 */}
-          <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-4">
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">
-                  總使用者數
-                </CardTitle>
-                <Users className="h-4 w-4 text-muted-foreground" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">
-                  {userStats.total_users || 0}
-                </div>
-                <p className="text-xs text-muted-foreground">系統註冊用戶</p>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">
-                  活躍使用者
-                </CardTitle>
-                <Users className="h-4 w-4 text-muted-foreground" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">
-                  {userStats.active_users || 0}
-                </div>
-                <p className="text-xs text-muted-foreground">最近30天登入</p>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">學生用戶</CardTitle>
-                <Users className="h-4 w-4 text-muted-foreground" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">
-                  {userStats.role_distribution?.student || 0}
-                </div>
-                <p className="text-xs text-muted-foreground">學生角色</p>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">本月新增</CardTitle>
-                <Users className="h-4 w-4 text-muted-foreground" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">
-                  {userStats.recent_registrations || 0}
-                </div>
-                <p className="text-xs text-muted-foreground">最近30天</p>
-              </CardContent>
-            </Card>
-          </div>
-
-          {/* 搜尋和篩選 */}
-          <Card className="border-nycu-blue-200">
-            <CardContent className="pt-4">
-              <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-                <div>
-                  <Label>搜尋使用者</Label>
-                  <Input
-                    placeholder="姓名、信箱或 NYCU ID"
-                    value={userSearch}
-                    onChange={e => setUserSearch(e.target.value)}
-                    className="border-nycu-blue-200"
-                  />
-                </div>
-                <div>
-                  <Label>角色篩選</Label>
-                  <select
-                    value={userRoleFilter}
-                    onChange={e => setUserRoleFilter(e.target.value)}
-                    className="w-full px-3 py-2 border border-nycu-blue-200 rounded-md"
-                  >
-                    <option value="">全部管理角色</option>
-                    <option value="super_admin">超級管理員</option>
-                    <option value="admin">管理員</option>
-                    <option value="college">學院</option>
-                    <option value="professor">教授</option>
-                  </select>
-                </div>
-                <div className="flex items-end gap-2">
-                  <Button
-                    onClick={handleSearch}
-                    className="flex-1 nycu-gradient text-white"
-                  >
-                    搜尋
-                  </Button>
-                  <Button
-                    onClick={clearFilters}
-                    variant="outline"
-                    className="border-nycu-blue-300 text-nycu-blue-600 hover:bg-nycu-blue-50"
-                  >
-                    清除
-                  </Button>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-
-          {/* 使用者列表 */}
-          <Card className="border-nycu-blue-200">
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Users className="h-5 w-5" />
-                使用者權限列表
-              </CardTitle>
-            </CardHeader>
-            <CardContent className="p-0">
-              {loadingUsers ? (
-                <div className="flex items-center justify-center py-8">
-                  <div className="flex items-center gap-3">
-                    <div className="animate-spin rounded-full h-6 w-6 border-2 border-nycu-blue-600 border-t-transparent"></div>
-                    <span className="text-nycu-navy-600">載入使用者中...</span>
-                  </div>
-                </div>
-              ) : usersError ? (
-                <div className="text-center py-12">
-                  <AlertCircle className="h-16 w-16 mx-auto mb-4 text-red-400" />
-                  <p className="text-lg font-medium text-red-600 mb-2">
-                    載入使用者失敗
-                  </p>
-                  <p className="text-sm text-gray-600 mb-4">{usersError}</p>
-                  <Button
-                    onClick={fetchUsers}
-                    variant="outline"
-                    className="border-red-300 text-red-600 hover:bg-red-50"
-                  >
-                    重試
-                  </Button>
-                </div>
-              ) : users.length > 0 ? (
-                <>
-                  <Table>
-                    <TableHeader>
-                      <TableRow>
-                        <TableHead className="font-bold px-5 py-3">
-                          使用者資訊
-                        </TableHead>
-                        <TableHead className="font-bold px-5 py-3">
-                          角色
-                        </TableHead>
-                        <TableHead className="font-bold px-5 py-3 w-40">
-                          單位
-                        </TableHead>
-                        <TableHead className="font-bold px-5 py-3">
-                          獎學金管理權限
-                        </TableHead>
-                        <TableHead className="font-bold px-5 py-3">
-                          註冊時間
-                        </TableHead>
-                        <TableHead className="font-bold px-5 py-3">
-                          最後登入
-                        </TableHead>
-                        <TableHead className="font-bold px-5 py-3">
-                          權限操作
-                        </TableHead>
-                      </TableRow>
-                    </TableHeader>
-                    <TableBody>
-                      {users.map(user => {
-                        const userPermissions = scholarshipPermissions.filter(
-                          p => p.user_id === Number(user.id)
-                        );
-                        return (
-                          <TableRow key={user.id}>
-                            <TableCell className="px-5 py-4 align-middle">
-                              <div className="space-y-1">
-                                <div className="font-medium">{user.name}</div>
-                                <div className="text-sm text-gray-500">
-                                  {user.email}
-                                </div>
-                                <div className="text-sm text-gray-500">
-                                  @{user.nycu_id}
-                                </div>
-                                {user.raw_data?.chinese_name && (
-                                  <div className="text-sm text-gray-500">
-                                    中文名: {user.raw_data.chinese_name}
-                                  </div>
-                                )}
-                                {user.raw_data?.english_name && (
-                                  <div className="text-sm text-gray-500">
-                                    英文名: {user.raw_data.english_name}
-                                  </div>
-                                )}
-                              </div>
-                            </TableCell>
-                            <TableCell className="px-5 py-4 align-middle">
-                              <Badge
-                                variant={
-                                  user.role === "super_admin"
-                                    ? "destructive"
-                                    : user.role === "admin"
-                                      ? "default"
-                                      : user.role === "college"
-                                        ? "secondary"
-                                        : user.role === "professor"
-                                          ? "outline"
-                                          : "default"
-                                }
-                                className="text-xs px-3 py-1 rounded-full whitespace-nowrap"
-                              >
-                                {getRoleLabel(user.role)}
-                              </Badge>
-                            </TableCell>
-                            <TableCell className="px-5 py-4 align-middle w-40">
-                              <div className="space-y-1">
-                                {user.dept_name ? (
-                                  <div className="text-sm font-medium text-gray-900 truncate">
-                                    {user.dept_name}
-                                  </div>
-                                ) : (
-                                  <div className="text-sm text-gray-400">
-                                    未設定
-                                  </div>
-                                )}
-                                {user.dept_code && (
-                                  <div className="text-xs text-gray-500">
-                                    代碼: {user.dept_code}
-                                  </div>
-                                )}
-                              </div>
-                            </TableCell>
-                            <TableCell className="px-5 py-4 align-middle">
-                              <div className="flex flex-wrap gap-2 min-h-[32px]">
-                                {loadingPermissions ? (
-                                  <div className="text-xs text-gray-400">
-                                    載入中...
-                                  </div>
-                                ) : user.role === "super_admin" ? (
-                                  <>
-                                    {availableScholarships.map(scholarship => (
-                                      <Badge
-                                        key={scholarship.id}
-                                        variant="default"
-                                        className="text-xs px-3 py-1 rounded-full mb-1"
-                                      >
-                                        {scholarship.name}
-                                      </Badge>
-                                    ))}
-                                    <div className="text-xs text-green-600 font-medium w-full">
-                                      擁有所有獎學金權限
-                                    </div>
-                                  </>
-                                ) : user.role === "professor" ? (
-                                  <div className="text-xs text-amber-600 font-medium">
-                                    教授無需管理權限
-                                  </div>
-                                ) : userPermissions.length === 0 ? (
-                                  <div className="text-xs text-gray-400">
-                                    無獎學金權限
-                                  </div>
-                                ) : (
-                                  userPermissions.map(permission => (
-                                    <Badge
-                                      key={permission.id}
-                                      variant="secondary"
-                                      className="text-xs px-3 py-1 rounded-full mb-1"
-                                    >
-                                      {permission.scholarship_name}
-                                    </Badge>
-                                  ))
-                                )}
-                              </div>
-                            </TableCell>
-                            <TableCell className="px-5 py-4 align-middle">
-                              <div className="text-sm text-gray-600">
-                                {new Date(user.created_at).toLocaleDateString(
-                                  "zh-TW",
-                                  {
-                                    year: "numeric",
-                                    month: "2-digit",
-                                    day: "2-digit",
-                                  }
-                                )}
-                              </div>
-                            </TableCell>
-                            <TableCell className="px-5 py-4 align-middle">
-                              <div className="text-sm text-gray-600">
-                                {user.last_login_at
-                                  ? new Date(user.last_login_at).toLocaleString(
-                                      "zh-TW",
-                                      {
-                                        year: "numeric",
-                                        month: "2-digit",
-                                        day: "2-digit",
-                                        hour: "2-digit",
-                                        minute: "2-digit",
-                                        second: "2-digit",
-                                        hour12: false,
-                                      }
-                                    )
-                                  : "從未登入"}
-                              </div>
-                            </TableCell>
-                            <TableCell className="px-5 py-4 align-middle">
-                              <div className="flex gap-1">
-                                <Button
-                                  variant="outline"
-                                  size="sm"
-                                  onClick={() => handleEditUser(user)}
-                                  className="hover:bg-nycu-blue-50 hover:border-nycu-blue-300"
-                                >
-                                  <Edit className="h-4 w-4" />
-                                  {user.role === "professor"
-                                    ? "更改角色"
-                                    : "編輯權限"}
-                                </Button>
-                              </div>
-                            </TableCell>
-                          </TableRow>
-                        );
-                      })}
-                    </TableBody>
-                  </Table>
-
-                  {/* 分頁 */}
-                  <div className="flex items-center justify-between p-4 border-t">
-                    <div className="text-sm text-gray-600">
-                      顯示 {(userPagination.page - 1) * userPagination.size + 1}{" "}
-                      到{" "}
-                      {Math.min(
-                        userPagination.page * userPagination.size,
-                        userPagination.total
-                      )}{" "}
-                      筆，共 {userPagination.total} 筆
-                    </div>
-                    <div className="flex gap-2">
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        onClick={() =>
-                          setUserPagination(prev => ({
-                            ...prev,
-                            page: prev.page - 1,
-                          }))
-                        }
-                        disabled={userPagination.page <= 1}
-                      >
-                        上一頁
-                      </Button>
-                      <span className="flex items-center px-3 text-sm">
-                        第 {userPagination.page} 頁
-                      </span>
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        onClick={() =>
-                          setUserPagination(prev => ({
-                            ...prev,
-                            page: prev.page + 1,
-                          }))
-                        }
-                        disabled={
-                          userPagination.page * userPagination.size >=
-                          userPagination.total
-                        }
-                      >
-                        下一頁
-                      </Button>
-                    </div>
-                  </div>
-                </>
-              ) : (
-                <div className="text-center py-12 text-gray-500">
-                  <Users className="h-16 w-16 mx-auto mb-4 text-gray-300" />
-                  <p className="text-lg font-medium">尚無使用者權限資料</p>
-                  <p className="text-sm mt-2 mb-4">
-                    點擊「新增使用者權限」開始設定使用者權限
-                  </p>
-                  <Button onClick={fetchUsers} variant="outline" size="sm">
-                    重新載入
-                  </Button>
-                </div>
-              )}
-            </CardContent>
-          </Card>
-
-          {/* 使用者編輯 Modal */}
-          <UserEditModal
-            isOpen={showUserForm}
-            onClose={() => setShowUserForm(false)}
-            editingUser={editingUser}
-            userForm={userForm}
-            onUserFormChange={handleUserFormChange}
-            onSubmit={handleUserSubmit}
-            isLoading={userFormLoading}
-            scholarshipPermissions={scholarshipPermissions}
-            availableScholarships={availableScholarships}
-            onPermissionChange={handlePermissionChange}
-          />
+          <UserPermissionManagement />
         </TabsContent>
 
         {hasQuotaPermission && (
